+++ directmemory-cache/src/main/java/org/apache/directmemory/memory/allocator/LazyUnsafeAllocator.java
@@ -108,7 +108,7 @@ public class LazyUnsafeAllocator
     }
 
     @Override
-    public int getCapacity()
+    public long getCapacity()
     {
         long capacity = 0;
         for ( UnsafePartitionBuffer partitionBuffer : partitionBuffers )
