+++ directmemory-cache/src/main/java/org/apache/directmemory/memory/allocator/MergingByteBufferAllocator.java
@@ -322,7 +322,7 @@ public class MergingByteBufferAllocator
     }
 
     @Override
-    public int getCapacity()
+    public long getCapacity()
     {
         return parentBuffer.capacity();
     }
